# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for work with metrics",
        "title": "SystemMetric API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/deleteMetric": {
            "delete": {
                "description": "delete metric with metricID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "delete page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metricID for deleting metric",
                        "name": "metricID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete metric is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/deleteMetricType": {
            "delete": {
                "description": "delete metric type with metricTypeID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "delete page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metricTypeID for deleting metric type",
                        "name": "metricTypeID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete metric type is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/getMetric": {
            "get": {
                "description": "get metric with metricID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metricID for getting metric",
                        "name": "metricID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get metric is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getMetricType": {
            "get": {
                "description": "get metricType with metricTypeID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "metricTypeID for getting metricType",
                        "name": "metricTypeID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get metricType is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/getMetricTypes": {
            "get": {
                "description": "get metric types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "get page",
                "responses": {
                    "200": {
                        "description": "Get metric types is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/insertMetric": {
            "post": {
                "description": "insert metric",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "insert page",
                "parameters": [
                    {
                        "description": "metric_name, timestamp, value, metric_type_id",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insert metric is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/insertMetricType": {
            "post": {
                "description": "insert metric type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "insert page",
                "parameters": [
                    {
                        "description": "type_name",
                        "name": "metric_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.MetricType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insert metric type is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/updateMetric": {
            "put": {
                "description": "update value with metricID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "update page",
                "parameters": [
                    {
                        "description": "metricID, value",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update metric is successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Metric": {
            "type": "object",
            "properties": {
                "metric_id": {
                    "type": "integer",
                    "example": 1
                },
                "metric_name": {
                    "type": "string",
                    "example": "metric name"
                },
                "metric_type_id": {
                    "type": "integer",
                    "example": 1
                },
                "timestamp": {
                    "type": "string",
                    "example": "1993-12-01 10:03:44.000 +0300"
                },
                "value": {
                    "type": "number",
                    "example": 0.5
                }
            }
        },
        "entity.MetricType": {
            "type": "object",
            "properties": {
                "type_id": {
                    "type": "integer",
                    "example": 1
                },
                "type_name": {
                    "type": "string",
                    "example": "type name"
                }
            }
        },
...
    }
}